#155: use deleteReply to consolidate useitem flow
